Gather all functions related to JSdoc annotations in one file, for JS components
