+++ packages/uxpin-merge-cli/src/steps/serialization/component/implementation/javascript/__tests__/getCommentTagInlineValue.test.ts
@@ -1,4 +1,4 @@
-import { getCommentTagInlineValue } from '../getCommentTagInlineValue';
+import { getCommentTagInlineValue } from '../jsdoc-uxpin-annotations';
 
 describe('getCommentTagInlineValue', () => {
   const tag = '@uxpintag';
