--- packages/uxpin-merge-cli/src/steps/serialization/component/implementation/javascript/getComponentDocUrlFromDescription.ts
@@ -1,16 +1,0 @@
-import { CommentTags } from '../../CommentTags';
-import { getCommentTagInlineValue } from './getCommentTagInlineValue';
-
-export function getComponentDocUrlFromDescription(description: string): string | undefined {
-  if (!description) {
-    return;
-  }
-
-  const componentDocUrl: string | undefined = getCommentTagInlineValue(description, CommentTags.UXPIN_DOC_URL);
-
-  if (!componentDocUrl) {
-    return;
-  }
-
-  return componentDocUrl;
-}
