--- packages/uxpin-merge-cli/src/steps/serialization/component/implementation/javascript/hasCommentTag.ts
@@ -1,3 +1,0 @@
-export function hasCommentTag(comment: string, tag: string): boolean {
-  return comment.includes(tag);
-}
